Expected Turnover Model/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
shap>=0.42.0
streamlit>=1.30.0
mplsoccer>=1.2.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
fastparquet>=2023.10.0